+++ src/Company/Company.php
@@ -34,6 +34,9 @@ final class Company
     /** @var string */
     private $description;
 
+    /** @var CompanyStatus */
+    private $status;
+
     /** @var string */
     private $zipcode;
 
@@ -102,6 +105,7 @@ final class Company
         $this->name = to_string($data['name']);
         $this->email = to_string($data['email']);
         $this->description = to_string($data['description']);
+        $this->status = isset($data['status']) ? new CompanyStatus(to_string($data['status'])) : null;
         $this->zipcode = to_string($data['zipcode']);
         $this->address = to_string($data['address']);
         $this->city = to_string($data['city']);
@@ -143,6 +147,11 @@ final class Company
     public function getDescription(): string
     {
         return $this->description;
+    }
+
+    public function getStatus(): ?CompanyStatus
+    {
+        return $this->status;
     }
 
     public function getZipcode(): string