+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Next release
+- Added `Wizaplace\SDK\Company\CompanyPatchCommand`
+- Added `Wizaplace\SDK\Company\CompanyStatus`
+- Added `Wizaplace\SDK\Company\Company::getStatus`
+- Added `Wizaplace\SDK\Company\CompanyService::patch`
+
+
 ## 1.101.0
 
 ## New features